Narrow InputSection.output_offset to u32 (96 -> 88 bytes)
output_offset is a subsection's offset within its output section. An
output section stays far below 4 GiB (a Mach-O image's addressable span
is bounded well under that), so a u32 holds it, and the "not yet
placed" marker used during thunk layout moves from u64::MAX to
u32::MAX. Absolute addresses (addr, chunk addr/fileoff) stay u64; the
arithmetic that combines them with output_offset casts it up at the use
site.

With is_alive already leaving padding in the struct, dropping this one
field crosses an alignment boundary: InputSection goes 96 -> 88 bytes,
another ~9% off the ~4.4M-element arena on a debug link. 322
integration tests pass and the linked clang and nu run.

+++ src/arch/arm64.rs
@@ -154,13 +154,13 @@ impl Arch for Arm64 {
                         continue 'hint;
                     };
                     let isec = &ctx.isecs[ctx.resolve_isec(isec)];
-                    if !isec.is_alive || isec.output_offset == u64::MAX {
+                    if !isec.is_alive || isec.output_offset == u32::MAX {
                         continue 'hint;
                     }
                     let chunk = &ctx.chunks[isec.osec as usize];
                     locs.push((
-                        (chunk.hdr.fileoff + isec.output_offset + off) as usize,
-                        chunk.hdr.addr + isec.output_offset + off,
+                        (chunk.hdr.fileoff + isec.output_offset as u64 + off) as usize,
+                        chunk.hdr.addr + isec.output_offset as u64 + off,
                     ));
                 }
                 let insn =
